% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls_analysis.R
\name{AUC_diff}
\alias{AUC_diff}
\title{Calculate the difference in Area-Under-the-Curve for DLS data}
\usage{
AUC_diff(data, treatments_in, control_in, metric)
}
\arguments{
\item{data}{A dataframe with DLS data. One row per sample}

\item{metric}{A string specifying the DLS metric used. Must be a substring of column names in 'data', excluding the 'sample' column.}

\item{treatments}{A vector of strings corresponding to a subset of the samples in 'data'.}

\item{control}{A string corresponding to one sample in 'data'.}
}
\description{
This function allows you to calculate the difference in AUC between treatment and control conditions.
}
\examples{
AUC_diff(dls_data, c('treatment_1', 'treatment_2'), 'control', 'intensities')
}
