% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls_analysis.R
\name{lm_test}
\alias{lm_test}
\title{Use linear model to predict titer loss based on a difference in AUC}
\usage{
lm_test(model, AUC)
}
\arguments{
\item{model}{A trained linear model. Output of lm_train().}

\item{AUC}{A dataframe with the difference in AUC for each treatment. Output of AUC_diff().}
}
\description{
This function allows you to predict titer loss based on a trained linear model and on a difference in AUC, which can be obtained with the function 'AUC_Diff'.
}
\examples{
lm_test(lm_model, AUC_data)
}
