% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls_analysis.R
\name{plot_DLS}
\alias{plot_DLS}
\title{Plot DLS curves}
\usage{
plot_DLS(data, treatments, control, metric = "Intensity", size)
}
\arguments{
\item{data}{A dataframe with DLS data. One row per sample}

\item{treatments}{A vector of strings corresponding to a subset of the samples in 'data'.}

\item{control}{A string corresponding to one sample in 'data'.}

\item{metric}{A character string specifying the DLS metric to use. Default is 'Intensity'. Other metrics typically include 'Volume' and 'Number'. Column names in 'data' must contain the substrings 'intens','volume' and 'number' respectively (excluding the 'sample' column).}

\item{size}{Optional. A vector of the particle sizes associated with the DLS measurements in 'data'.}
}
\description{
This function allows you to plot different DLS curves on the same graph for comparison.
}
\examples{
plot_DLS(dls_data, c('treatment_1', 'treatment_2'), 'control', 'Intensity', size_vector)
}
