% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls_analysis.R
\name{plot_model}
\alias{plot_model}
\title{Plot linear model for titer loss over difference in AUC.}
\usage{
plot_model(model)
}
\arguments{
\item{model}{A trained linear model. Output of lm_train.}
}
\description{
This function allows you to plot the training data and the trained linear model of the titer loss
associated with the difference in AUC calculated from DLS data.
}
\examples{
plot_model(AUC_data, titer_data, 'control')
}
