% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls_analysis.R
\name{plot_test_data}
\alias{plot_test_data}
\title{Plot titer loss predictions}
\usage{
plot_test_data(model, AUC)
}
\arguments{
\item{model}{A linear model.}

\item{AUC}{A dataframe with the difference in AUC for each treatment.}
}
\description{
This function plots titer loss predictions with the training dat and model
}
\examples{
plot_test_data(lm_model, AUC_data)
}
