% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dls_analysis.R
\name{run_all}
\alias{run_all}
\title{Process DLS and titer training data, train model and predict titer loss for new DLS data}
\usage{
run_all(
  data,
  lytic,
  training_pairs,
  testing_pairs,
  metric = "Intensity",
  intercept = TRUE
)
}
\arguments{
\item{data}{A dataframe with DLS data. One row per sample}

\item{lytic}{A dataframe with the titer associated with each treatment.}

\item{training_pairs}{A dataframe with two columns, named 'treatment' and 'control'.}

\item{testing_pairs}{Optional. A dataframe with two columns, named 'treatment' and 'control'.}

\item{metric}{A character string specifying the DLS metric to use. Default is 'Intensity'. Other metrics typically include 'Volume' and 'Number'. Column names in 'data' must contain the substrings 'intens','volume' and 'number' respectively (excluding the 'sample' column).}

\item{intercept}{If TRUE, include a non-zero intercept in the linear model. Default is TRUE.}
}
\description{
This is a wrapper function that runs all of the analysis, replicating results from the Phage-ELF shiny app.
}
\examples{
run_all(dls_data, lytic_data, training_pairs, testing pairs)
}
